+++ overwin/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from.models import *
+from.models import User,GamePlayer,FavoriteGamePlayer,Recruitment
 
 admin.site.register(User)
 admin.site.register(GamePlayer)
