+++ overwin/forms/account_register_and_login.py
@@ -1,0 +1,7 @@
+from django.contrib.auth.forms import UserCreationForm
+from .models import User
+
+class UserRegisterForm(UserCreationForm):
+    class Meta:
+        model = User
+        fields = ('username', 'email')
