+++ overwin/forms/party_recruitment.py
@@ -1,0 +1,7 @@
+from .models import Recruitment
+from django import forms
+
+class RecruitmentForm(forms.ModelForm):
+    class Meta:
+        model = Recruitment
+        fields = ['max_recruit_member', 'comment']
