+++ overwin/views/account_info.py
@@ -1,5 +1,5 @@
-from ..models import *
-from ..forms import *
+from ..models import User
+from ..forms import UserRegisterForm
 from django.shortcuts import render, get_object_or_404,redirect
 from django.contrib.auth.decorators import login_required
 from django.urls import reverse
